Move matrix chain multiplication methods into a separate file under divide and conquer module
