+++ ruby/divide_and_conquer/divide_and_conquer_tests.sh
@@ -1,1 +1,2 @@
 ruby unit_tests/find_max_subarray_test.rb
+ruby unit_tests/matrix_multiplication_test.rb
